TF: Fix windows build (implicit cast between double and int32_t).
R=bmeurer@chromium.org
BUG=

Review URL: https://codereview.chromium.org/431963002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@22754 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ test/cctest/compiler/test-run-machops.cc
@@ -2945,14 +2945,14 @@ TEST(RunConvertFloat64ToInt32_B) {
   for (int32_t n = 1; n < 31; ++n) {
     {
       input = 1 << n;
-      int32_t expect = input;
+      int32_t expect = static_cast<int32_t>(input);
       CHECK_EQ(expect, m.Call());
       CHECK_EQ(expect, output);
     }
 
     {
       input = 3 << n;
-      int32_t expect = input;
+      int32_t expect = static_cast<int32_t>(input);
       CHECK_EQ(expect, m.Call());
       CHECK_EQ(expect, output);
     }
